Cleanups in all analytics plugins.
BREAKING CHANGES

- Classes moved to package com.kaltura.playkit.plugins.ott:
	PhoenixAnalyticsPlugin, PhoenixAnalyticsEvent, PhoenixAnalyticsConfig,
	TVPAPIAnalyticsPlugin, TVPAPIAnalyticsEvent, TVPAPIAnalyticsConfig

- Classes moved to package com.kaltura.playkit.plugins.ovp:
	KalturaAnalyticsPlugin, KalturaLiveStatsConfig, KalturaLiveStatsEvent, KalturaLiveStatsPlugin, KalturaStatsConfig, KalturaStatsEvent, KalturaStatsPlugin
