+++ playkit/src/androidTest/java/com/kaltura/playkit/plugins/ovp/KalturaAnalyticsAndroidTest.java
@@ -1,4 +1,4 @@
-package com.kaltura.playkit.plugins;
+package com.kaltura.playkit.plugins.ovp;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
@@ -11,8 +11,10 @@ import com.kaltura.playkit.PKEvent;
 import com.kaltura.playkit.PKMediaConfig;
 import com.kaltura.playkit.PKMediaEntry;
 import com.kaltura.playkit.PKMediaSource;
+import com.kaltura.playkit.PlayKitManager;
 import com.kaltura.playkit.Player;
 import com.kaltura.playkit.PlayerEvent;
+import com.kaltura.playkit.plugins.MockPlayer;
 
 import org.junit.Assert;
 import org.junit.Before;
@@ -22,25 +24,22 @@ import org.junit.runner.RunWith;
 import java.util.ArrayList;
 
 /**
- * Created by zivilan on 11/12/2016.
+ * Created by zivilan on 12/12/2016.
  */
-
 @RunWith(AndroidJUnit4.class)
-public class KalturaStatsAndroidTest {
+public class KalturaAnalyticsAndroidTest {
 
     private Player player;
     private Context context;
     private JsonObject pluginConfig;
     private MessageBus messageBus;
-    private KalturaStatsPlugin plugin;
+    private KalturaAnalyticsPlugin plugin;
     private int duration = 3000;
     private long seek = 100;
     private String entryId = "259295";
     private int partnerId = 1281471;
-    private String widgetId = "_" + partnerId;
     private int uiconfId = 24997472;
     private PlayerEvent event;
-    private boolean isSeek;
 
     @Before
     public void setUp(){
@@ -53,7 +52,7 @@ public class KalturaStatsAndroidTest {
         ((MockPlayer) player).setDuration(duration);
         player.seekTo(seek);
 
-        plugin = (KalturaStatsPlugin) KalturaStatsPlugin.factory.newInstance();
+        plugin = (KalturaAnalyticsPlugin) KalturaAnalyticsPlugin.factory.newInstance();
         plugin.onLoad(player, pluginConfig, messageBus, context);
     }
 
@@ -62,7 +61,7 @@ public class KalturaStatsAndroidTest {
         pluginConfig = new JsonObject();
         pluginConfig.addProperty("sessionId", "b3460681-b994-6fad-cd8b-f0b65736e837");
         pluginConfig.addProperty("uiconfId", uiconfId);
-        pluginConfig.addProperty("baseUrl", "https://stats.kaltura.com/api_v3/index.php");
+        pluginConfig.addProperty("baseUrl", "https://analytics.kaltura.com/api_v3/index.php");
         pluginConfig.addProperty("partnerId", partnerId);
         pluginConfig.addProperty("timerInterval", 30);
     }
@@ -80,20 +79,19 @@ public class KalturaStatsAndroidTest {
     @Test
     public void testSeekPlugin(){
         event = new PlayerEvent(PlayerEvent.Type.SEEKED);
-        isSeek = true;
         messageBus.listen(new PKEvent.Listener() {
             @Override
             public void onEvent(PKEvent event) {
-                Assert.assertTrue(((LogEvent)event).log.contains(KalturaStatsPlugin.factory.getName()));
+                Assert.assertTrue(((LogEvent)event).log.contains(KalturaAnalyticsPlugin.factory.getName()));
                 Assert.assertTrue(((LogEvent)event).log.contains("SEEK"));
-                Assert.assertTrue(((LogEvent)event).request.contains("duration=" + (duration / 1000)));
-                Assert.assertTrue(((LogEvent)event).request.contains("eventType=" + KalturaStatsPlugin.KStatsEvent.SEEK.getValue()));
-                Assert.assertTrue(((LogEvent)event).request.contains("currentPoint=" + seek));
-                Assert.assertTrue(((LogEvent)event).request.contains("seek=" + isSeek));
+                Assert.assertTrue(((LogEvent)event).request.contains("action=trackEvent"));
+                Assert.assertTrue(((LogEvent)event).request.contains("eventType=" + KalturaAnalyticsPlugin.KAnalonyEvents.SEEK.getValue()));
+                Assert.assertTrue(((LogEvent)event).request.contains("playbackType=live"));
+                Assert.assertTrue(((LogEvent)event).request.contains("clientVer=" + PlayKitManager.CLIENT_TAG));
                 Assert.assertTrue(((LogEvent)event).request.contains("entryId=" + entryId));
                 Assert.assertTrue(((LogEvent)event).request.contains("uiconfId=" + uiconfId));
-                Assert.assertTrue(((LogEvent)event).request.contains("widgetId=" + widgetId));
-                Assert.assertTrue(((LogEvent)event).request.contains("partnerId=" + partnerId));
+                Assert.assertTrue(((LogEvent)event).request.contains("position=" + seek));
+                Assert.assertTrue(((LogEvent)event).request.contains("eventIndex=" + 0));
             }
         }, LogEvent.LogType.LogEvent);
         messageBus.post(event);
@@ -102,20 +100,19 @@ public class KalturaStatsAndroidTest {
     @Test
     public void testPlayPlugin(){
         event = new PlayerEvent(PlayerEvent.Type.PLAY);
-        isSeek = false;
         messageBus.listen(new PKEvent.Listener() {
             @Override
             public void onEvent(PKEvent event) {
-                Assert.assertTrue(((LogEvent)event).log.contains(KalturaStatsPlugin.factory.getName()));
+                Assert.assertTrue(((LogEvent)event).log.contains(KalturaAnalyticsPlugin.factory.getName()));
                 Assert.assertTrue(((LogEvent)event).log.contains("PLAY"));
-                Assert.assertTrue(((LogEvent)event).request.contains("duration=" + (duration / 1000)));
-                Assert.assertTrue(((LogEvent)event).request.contains("eventType=" + KalturaStatsPlugin.KStatsEvent.PLAY.getValue()));
-                Assert.assertTrue(((LogEvent)event).request.contains("currentPoint=" + seek));
-                Assert.assertTrue(((LogEvent)event).request.contains("seek=" + isSeek));
+                Assert.assertTrue(((LogEvent)event).request.contains("action=trackEvent"));
+                Assert.assertTrue(((LogEvent)event).request.contains("eventType=" + KalturaAnalyticsPlugin.KAnalonyEvents.PLAY_REQUEST.getValue()));
+                Assert.assertTrue(((LogEvent)event).request.contains("playbackType=live"));
+                Assert.assertTrue(((LogEvent)event).request.contains("clientVer=" + PlayKitManager.CLIENT_TAG));
                 Assert.assertTrue(((LogEvent)event).request.contains("entryId=" + entryId));
                 Assert.assertTrue(((LogEvent)event).request.contains("uiconfId=" + uiconfId));
-                Assert.assertTrue(((LogEvent)event).request.contains("widgetId=" + widgetId));
-                Assert.assertTrue(((LogEvent)event).request.contains("partnerId=" + partnerId));
+                Assert.assertTrue(((LogEvent)event).request.contains("position=" + seek));
+                Assert.assertTrue(((LogEvent)event).request.contains("eventIndex=" + 0));
             }
         }, LogEvent.LogType.LogEvent);
         messageBus.post(event);
