+++ playkit/src/main/java/com/kaltura/playkit/plugins/ovp/KalturaLiveStatsConfig.java
@@ -1,4 +1,4 @@
-package com.kaltura.playkit.plugins.configs;
+package com.kaltura.playkit.plugins.ovp;
 
 import com.google.gson.JsonObject;
