+++ playkit/src/main/java/com/kaltura/playkit/plugins/ovp/KalturaLiveStatsPlugin.java
@@ -1,4 +1,4 @@
-package com.kaltura.playkit.plugins;
+package com.kaltura.playkit.plugins.ovp;
 
 import android.content.Context;
 
@@ -19,7 +19,6 @@ import com.kaltura.playkit.PlaybackInfo;
 import com.kaltura.playkit.Player;
 import com.kaltura.playkit.PlayerEvent;
 import com.kaltura.playkit.api.ovp.services.LiveStatsService;
-import com.kaltura.playkit.plugins.configs.KalturaLiveStatsConfig;
 import com.kaltura.playkit.utils.Consts;
 
 import java.util.Date;
